refactored notify function into module
